% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_setup.R
\name{git_ignores}
\alias{git_ignores}
\title{git_ignores for usethis::use_git_ignore()}
\usage{
git_ignores()
}
\description{
git_ignores for usethis::use_git_ignore()
}
\examples{
\dontrun{
usethis::use_git_ignore(ignores = git_ignores())
}

}
