% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_tools.R
\name{theme_blog}
\alias{theme_blog}
\title{ggplot2 themes}
\usage{
theme_blog(...)
}
\arguments{
\item{...}{passed to ggplot2 theme()}
}
\description{
ggplot2 themes
}
\details{
Themes all tweak existing themes but accept any ggplot2 theme arguments to overwrite settings
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "My Plot") +
    theme_blog(text = element_text(size = 20))
}
}
